rearranged firebase and model stuff out into their own respective files to make the store more manageable
