+++ permission-portal-frontend/client/src/store/index.js
@@ -1,228 +1,8 @@
 import React from 'react'
-import app from 'firebase/app'
-import 'firebase/auth'
-import 'firebase/firestore'
-import 'firebase/database'
-import 'firebase/functions'
-import * as firebaseConfigLocal from '../config/firebase.config.local'
-import * as firebaseConfigDev from '../config/firebase.config.dev'
-import * as firebaseConfigTest from '../config/firebase.config.test'
-import * as firebaseConfigProd from '../config/firebase.config.prod'
-import * as firebaseConfigStaging from '../config/firebase.config.staging'
-import { types, cast, flow, onSnapshot } from 'mobx-state-tree'
-
-var firebaseConfigMap = {
-  development: firebaseConfigDev,
-  test: firebaseConfigTest,
-  production: firebaseConfigProd,
-  local: firebaseConfigLocal,
-  staging: firebaseConfigStaging,
-}
-
-const config = firebaseConfigMap[process.env ? process.env.NODE_ENV : 'development']
-
-console.log(`process.env.NODE_ENV = ${process.env.NODE_ENV}`)
-
-app.initializeApp(config)
-const auth = app.auth()
-const db = app.firestore()
-const createUserCallable = app.functions().httpsCallable('createUser')
+import { rootStore, defaultUser, defaultOrganization } from './model'
+import { auth, db } from './firebase'
 
 const RootStoreContext = React.createContext()
-
-const User = types
-  .model({
-    isSignedIn: types.maybe(types.boolean),
-    email: types.string,
-    isAdmin: types.boolean,
-    isSuperAdmin: types.boolean,
-    disabled: types.boolean,
-    prefix: types.maybe(types.string),
-    imageBlob: types.maybeNull(types.string),
-    firstName: types.string,
-    lastName: types.string,
-    organizationID: types.string,
-  })
-  .actions((self) => {
-    const __update = (updates) => {
-      Object.keys(updates).forEach((key) => {
-        if (self.hasOwnProperty(key)) self[key] = updates[key] // eslint-disable-line no-prototype-builtins
-      })
-      console.log('Updated User:')
-      console.log(self)
-    }
-    const update = flow(function* (updates) {
-      try {
-        yield db.collection('users').doc(self.email).update(updates)
-      } catch (err) {
-        console.error('Error updating users', err)
-      }
-    })
-    const updateImage = flow(function* (blob) {
-      try {
-        // .set() with { merge: true } so that if the document dne, it's created, otherwise its updated
-        yield db.collection('userImages').doc(self.email).set({ blob: blob }, { merge: true })
-      } catch (err) {
-        console.error('Error updating image', err)
-      }
-    })
-
-    return { __update, update, updateImage }
-  })
-
-const Organization = types
-  .model({
-    id: types.string,
-    name: types.string,
-    welcomeText: types.string,
-    enableExposureText: types.string,
-    recommendExposureText: types.string,
-    notifyingOthersText: types.string,
-    exposureInfoText: types.string,
-    exposureAboutText: types.string,
-    exposureDetailsText: types.string,
-    exposureDetailsLearnText: types.string,
-    verificationStartText: types.string,
-    verificationIdentifierText: types.string,
-    verificationIdentifierAboutText: types.string,
-    verificationAdministrationDateText: types.string,
-    verificationReviewText: types.string,
-    verificationSharedText: types.string,
-    verificationNotSharedText: types.string,
-    diagnosisText: types.string,
-    exposureText: types.string,
-    members: types.array(User),
-  })
-  .actions((self) => {
-    const __update = (updates) => {
-      Object.keys(updates).forEach((key) => {
-        if (self.hasOwnProperty(key)) self[key] = updates[key] // eslint-disable-line no-prototype-builtins
-      })
-      console.log('Updated Organization:')
-      console.log(self)
-    }
-
-    const update = flow(function* (updates) {
-      try {
-        yield db.collection('organizations').doc(self.id).update(updates)
-      } catch (err) {
-        console.error('Error updating organization texts', err)
-        throw err
-      }
-    })
-
-    const __setMembers = (members) => {
-      self.members = cast(members)
-      console.log('Set members:')
-      console.log(self.members[0])
-    }
-
-    return { __update, __setMembers, update }
-  })
-
-const Store = types
-  .model({
-    user: User,
-    organization: Organization,
-  })
-  .actions(() => {
-    const signInWithEmailAndPassword = flow(function* (email, password) {
-      yield auth.signInWithEmailAndPassword(email, password)
-    })
-
-    const signOut = flow(function* () {
-      yield auth.signOut()
-    })
-
-    const createUser = flow(function* (newUser) {
-      try {
-        const result = yield createUserCallable(newUser)
-        console.log(`Created new user: ${JSON.stringify(result.data)}`)
-        return result.data
-      } catch (err) {
-        console.log(err)
-        throw err
-      }
-    })
-
-    const sendPasswordResetEmail = flow(function* (email) {
-      try {
-        yield auth.sendPasswordResetEmail(email)
-        return true
-      } catch (err) {
-        console.warn(err)
-        throw err
-      }
-    })
-
-    const updateUserByEmail = flow(function* (email, updates) {
-      try {
-        yield db.collection('users').doc(email).update(updates)
-      } catch (err) {
-        console.error(`Error updating user: ${email}`, err)
-        throw err
-      }
-    })
-
-    return { signInWithEmailAndPassword, signOut, createUser, sendPasswordResetEmail, updateUserByEmail }
-  })
-
-const defaultUser = {
-  isSignedIn: false,
-  email: '',
-  isAdmin: false,
-  isSuperAdmin: false,
-  disabled: false,
-  prefix: '',
-  firstName: '',
-  lastName: '',
-  organizationID: '',
-  imageBlob: null,
-}
-
-const defaultOrganization = {
-  id: '',
-  name: '',
-  welcomeText: '',
-  enableExposureText: '',
-  recommendExposureText: '',
-  notifyingOthersText: '',
-  exposureInfoText: '',
-  exposureAboutText: '',
-  exposureDetailsText: '',
-  exposureDetailsLearnText: '',
-  verificationStartText: '',
-  verificationIdentifierText: '',
-  verificationIdentifierAboutText: '',
-  verificationAdministrationDateText: '',
-  verificationReviewText: '',
-  verificationSharedText: '',
-  verificationNotSharedText: '',
-  diagnosisText: '',
-  exposureText: '',
-  members: [],
-}
-
-const defaultStore = {
-  user: defaultUser,
-  organization: defaultOrganization,
-}
-
-let initialStore = defaultStore
-
-// Based on https://egghead.io/lessons/react-store-store-in-local-storage
-if (localStorage.getItem('store')) {
-  initialStore = JSON.parse(localStorage.getItem('store'))
-}
-
-const rootStore = Store.create({
-  ...initialStore,
-})
-
-// Based on https://egghead.io/lessons/react-store-store-in-local-storage
-onSnapshot(rootStore, (snapshot) => {
-  localStorage.setItem('store', JSON.stringify(snapshot))
-})
 
 const createStore = (WrappedComponent) => {
   return class extends React.Component {
